Remove dependency from config to utils
